Added basic modelling
In this commit I started creating the basic modelling for the database.
So far it is just a really simple graph network.

I updated the postgres image along the way - had some issues bringing up
the project and though a newer image might fix it. It didn't, the error
was somewhere else. But updates are a good thing, right?

Added more dependencies for making SQLAlchemy work with a Postgres
database, as well as the UUID package for being able to use uuids rather
than integer based ones.

Added a small file that holds values and / or retrieves values from the
environment, this way things like the db connection are configurable
form the outside.

And most importantly for the next steps, I added a `__main__.py` file to
run when the package is called directly. In this the database gets
initialised. And in the next commits I'll start with addind data to the
database there.