Adding 'writings' to the database
With this commit I added code for inserting all 'writings' into the
database. With 'writings' I am referring to what is called 'keb' in
the JMdict XML, or all words there are in the Japanese language as they
are written.

I needed to update the model definitions as I was running into
SQL-ALchemy errors, redefining the relationships with `backref`s rather
than `back_populates`, a definition on how to join things up, and
defaults for the uuids.

I also added the full JMdict XML file as I couldn't find it as a package
or something that is under version control, otherwise I would have jsut
downloaded it in the code or Docker build.

Furthermore I exposed the postges containers port so that I can verify
the databases content from my machine.