Writings fallback to `r_ele` when there is no `k_ele`
In the last commit I added code to parse the JMdict and add all writings
there are in it. With writings meaning words from the JMdict that can
appear in written form in Japanese media. Basically these are all words
there are in the Japanese language in the form you write them.

The problem was, that I only relied on the `k_ele` element for sourcing
those words. This, however, excluded all words that have no kanji
representation, meaning they are entirely and only written in kana.

So with this commit I added a fallback to the parser. Whenever there is
no `k_ele` we fall back to using the `r_ele` as a writing.
I removed the condition for filtering out outdated Kanji use from the
`k_ele` xpath, as I would not want to filter out those words - the
information about those words using outdated kanji can be added later -
and also it made no difference. No words were added after I removed the
condition.

Perhaps I should rename 'writing' into 'word' or something. Writing
seems a bit ambiguous.