Reworked writings insert
In this commit I reworked the way how writings are inserted into the
database. I did this in preparation to also add readings in the next few
commits. What I did here is develop a technique that would allow me to
insert writings and readings in a similar fashion, with little code
duplication and without having troubles with existing content in the
database.