Installed arango db
In this commit I installed ArangoDB in the project. That means I removed
all SQL-Alchemy code that was not connected to reading the JMDict,
removed SQL-Alchemy from the `setup.py` and from the
`docker-compose.yml`. I added configuration for using the ArangoDB image
in the `docker-compose.yml`, added the `arango-orm` package to our
`setup.py`, adapted the `dbURL` and renamed it `dbHostURL`, and finally
added code to establish a connection to the database.
As all the other code, that extracts the data from the JMDict is not
adapted yet, I just commented out the calls to those parts. In the next
commits I'll add some graph definitions and adapt the JMDict code to
work with the new database / orm wrapper.